Initial commit: sets up basic TCP echo server
This has two implementations, which were done as mental exercises

The currently used implementation looks for newline characters, which
may not be reasonable in the future depending on what type of data is
being communicated. But it works fine for now.

The unused implementation is protocol-agnostic, and will simply pass on
bytes. This also uses a fixed buffer which can prevent a buggy or
malicious client from sending a lot of data which is allocated when
using 'ReadBytes'. But this implementation is more verbose and may
mangle messages depending on exactly how a "message" is defined

Keeping both for now so I can figure out which makes more sense as new
features are added