Disconnects clients stop sending data for too long
To defend against slowloris-style behavior

We peek at the first byte that is received - that's the trigger for us
to know the client is starting to be active. Totally inactive clients
that never send data aren't affected. When the client sends that first
byte, we aggregate the following bytes in a buffer while looking for a
newline.